runtime: remove one last float64

+++ src/runtime/mprof.go
@@ -279,7 +279,8 @@ func SetBlockProfileRate(rate int) {
 		r = 1 // profile everything
 	} else {
 		// convert ns to cycles, use float64 to prevent overflow during multiplication
-		r = int64(float64(rate) * float64(tickspersecond()) / (1000 * 1000 * 1000))
+		// TODO(aram): re-add float64 here.
+		r = int64(rate) * tickspersecond() / (1000 * 1000 * 1000)
 		if r == 0 {
 			r = 1
 		}
